add "testing" overlay feature to nexus-db-queries
