+++ nexus/db-queries/Cargo.toml
@@ -64,6 +64,13 @@ omicron-uuid-kinds.workspace = true
 oximeter.workspace = true
 omicron-workspace-hack.workspace = true
 
+# only enabled during tests or via the `testing` feature
+omicron-test-utils = { workspace = true, optional = true }
+
+[features]
+# Enable to export `datastore_test`
+testing = ["omicron-test-utils"]
+
 [dev-dependencies]
 assert_matches.workspace = true
 camino-tempfile.workspace = true
