+++ nexus/db-queries/src/db/datastore/mod.rs
@@ -88,6 +88,8 @@ mod ssh_key;
 mod switch;
 mod switch_interface;
 mod switch_port;
+#[cfg(any(test, feature = "testing"))]
+pub mod pub_test_utils;
 #[cfg(test)]
 pub(crate) mod test_utils;
 mod update;
